pyproject.toml
[tool.poetry]
name = "freechat"
version = "0.1.0"
description = "Example LangGraph project for deployment to LangGraph Cloud"
authors = [
    "langchain-ai"
]
packages = [
    { include = "freechat" },
]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
langgraph = "^0.2.0"
langchain_anthropic = "^0.1.0"
langchain_core = "^0.2.0"
langchain_openai = "^0.1.0"
tavily-python = "^0.3.0"
langchain_community = "^0.2.0"
langgraph-cli = "^0.1.5"
python-dotenv = "^0.19.1"
langchain-pinecone = "*"
pinecone-client = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"